+++ drivers/cam_sensor_module/cam_csiphy/cam_csiphy_core.h
@@ -1,7 +1,7 @@
 /* SPDX-License-Identifier: GPL-2.0-only */
 /*
  * Copyright (c) 2017-2018, 2020-2021 The Linux Foundation. All rights reserved.
- * Copyright (c) 2022 Qualcomm Innovation Center, Inc. All rights reserved.
+ * Copyright (c) 2022-2023 Qualcomm Innovation Center, Inc. All rights reserved.
  */
 
 #ifndef _CAM_CSIPHY_CORE_H_
@@ -91,5 +91,5 @@ int cam_csiphy_dump_status_reg(struct csiphy_device *csiphy_dev);
  * correct PHY hw.
  *
  */
-int cam_csiphy_util_update_aon_registration(uint32_t phy_idx, uint8_t aon_camera_id);
+int cam_csiphy_util_update_aon_registration(uint32_t phy_idx, uint32_t aon_camera_id);
 #endif /* _CAM_CSIPHY_CORE_H_ */
